+++ tests/test_layout.py
@@ -76,3 +76,17 @@ def test_tree():
     expected = "⬍ Layout(name='root')                                       \n├── ⬍ Layout(size=2)                                        \n└── ⬌ Layout(name='bar')                                    \n    ├── ⬍ Layout()                                          \n    └── ⬍ Layout()                                          \n"
 
     assert result == expected
+
+
+def test_refresh_screen():
+    layout = Layout()
+    layout.split_row(Layout(name="foo"), Layout(name="bar"))
+    console = Console(force_terminal=True, width=20, height=5)
+    console.print(layout)
+    with console.screen():
+        with console.capture() as capture:
+            layout.refresh_screen(console, "foo")
+    result = capture.get()
+    print(repr(result))
+    expected = "\x1b[1;1H\x1b[34m╭─\x1b[0m\x1b[34m \x1b[0m\x1b[32m'foo'\x1b[0m\x1b[34m─╮\x1b[0m\x1b[2;1H\x1b[34m│\x1b[0m Layout \x1b[34m│\x1b[0m\x1b[3;1H\x1b[34m│\x1b[0m \x1b[1m(\x1b[0m      \x1b[34m│\x1b[0m\x1b[4;1H\x1b[34m│\x1b[0m     \x1b[33mna\x1b[0m \x1b[34m│\x1b[0m\x1b[5;1H\x1b[34m╰────────╯\x1b[0m"
+    assert result == expected